pico: sd card: Update DMA and transfer speed
It seems that the DMA does not coexist nicely between the SD code
and the i2s audio code.  This could be either package's fault or
it could just be that it causes missed interrupts due to the fact
that both are triggering constantly.

Either way, move the SD to dma 1 and leave the audio on dma 0 to
fix problems where playing audio from the sdcard sometimes freezes.

With that seemingly fixed, also restore the original transfer
speed of the SD card.  I had previously lowered it assuming that too
high a speed was causing the SD card to hang when in fact it seems
like it was most likely the conflict.