audio: pico: Replace pico_audio_i2s with my own version
There seems to be a bug in the pico-extras library that causes it
to periodically hang and "crash" the program.  I susepct it is
related to the interaction between the threading and however they
are trying to pass the buffers around.

Either way, this code is not particularly nice and doesn't integrate
well with FreeRTOS so it makes sense to replace it.

This version uses the NOTIFY mechanism of FreeRTOS to easily coordinate
between threads trying to play audio and the IRQ handler.