Build system redo
* Make a single library lib-pi which contains everything
* Make sure we don't build all the pico / freertos / fat code for
  everything that touches the threads
* Make sure that we don't leak any freertos / fat / alsa internals
  out of this library

This speeds up the pico build approximately 5x.

This changes all the .uf2 file sizes.  Everything that includes the
threads is a decent amount smaller and some things that are small
got a little bigger.  Overall a file size win.

This also ensures that we aren't building things like the core
networking libraries with different setups and then trying to link
those 2 builds together (or freertos or ...).

This will have killed the ability to specify each executable's
hostname for the network.