net: lwipopts.h: Go through all the options
Aside from the debugging options (which are useful to have if you
ever have to enable debugging), this file now contains only changes
that I personally think might be needed by our code and doesn't
reproduce a random subset of all the default options.

Note: This actually solved the problem of having the server be single
threaded.