complete build redo as INTERFACE libraries
I just lost a week or more of time trying to figure out a strange
crash in the bluetooth flash writing.  It turns out that the lib-pi
build was using threads causing it to appear like pico_flash was compiled
with threaded support when in reality one the pico-sdk libraries included
in the executable build was including pico_flash which caused the actual
inmplementation not to be built with threaded support causing it to crash.

This is strong evidence that trying to fight the good fight to encapsulate
the pico implementation in the library is a losing battle.

Switch to building everything as interface libraries and bite the cost
of recompiling everything everywhere.

To mitigate this make the examples (previously utils) optionally built.