pico: add in hooks for irq setting when multithreaded
The setting of a gpio's irq handler is not locked by the system
and if you run it without any special handling, it can fail due
to changes cores in the middle of the sdk call.

Handle this by letting the thread system set a hook which sets
core affinity to core 0 before making the irq call and then restores
the affinity afterward.

I choose core 0 because the async context (and likely other background
threads) use core 1. This lets us balance out the work a little bit.