pico: get rid of the long press bootsel to reboot bootloader
This doesn't work on pico2 and I really don't need it.  For active
development, I will be using a pico probe which lets me flash
directly.  For field debugging/fixing, I'm going to need to plug
in a usb cable anyway and shouldn't be doing many iterations of it
anyway...