pngle: move png code to this library
Which lets us use png's on the pico.  This lacks error handling
which needs to be added along with the ability to stream directly
from the file to the display to support larger displays.

Unify the all display examples into a single examples that you
can change the display to test with.